Update Assignment-1 Solutions.cs

+++ Assignment-1 Solutions.cs
@@ -8,7 +8,7 @@ namespace ConsoleApplication
         static void Main(string[] args)
         {
             Console.WriteLine("Enter the ph level number of the Aquarium");
-            int ph = Convert.ToInt32(Console.ReadLine());
+            double ph = Convert.ToInt32(Console.ReadLine());
 
             if (ph == 7)
             {
